requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
GDAL @ file:///D:/geoinformatyka/SEMESTRVI/seminarium_inzynierskie/GDAL-3.4.2-cp39-cp39-win_amd64.whl
haversine==2.7.0
idna==3.3
nodeenv==1.7.0
numpy==1.23.4
psycopg2==2.9.3
PyJWT==2.6.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11